image-segmentation/requirements.txt
torch
torchvision
pydicom
numpy
pillow
tqdm